refactor: make convertDpToPixel as top level
